requirements.txt
apturl==0.5.2
asn1crypto==0.22.0
beautifulsoup4==4.4.1
blinker==1.3
boto==2.48.0
Brlapi==0.6.4
bz2file==0.98
cffi==1.10.0
chardet==2.3.0
checkbox-support==0.22
command-not-found==0.3
cryptography==1.2.3
cssselect==1.0.1
cycler==0.10.0
defer==1.0.6
Django==1.11.3
feedparser==5.1.3
gensim==3.1.0
gmail==0.0.5
guacamole==0.9.2
html5lib==0.999
httplib2==0.9.1
idna==2.5
Jinja2==2.8
language-selector==0.1
louis==2.6.4
lxml==3.5.0
Mako==1.0.3
MarkupSafe==0.23
matplotlib==2.1.2
nltk==3.2.4
numpy==1.13.3
oauthlib==1.0.3
onboard==1.2.0
padme==1.1.1
pexpect==4.0.1
pickle-warehouse==0.1.1
Pillow==3.1.2
plainbox==0.25
psutil==5.4.5
ptyprocess==0.5
py-stackexchange==2.2.7
pyasn1==0.1.9
pycorenlp==0.3.0
pycparser==2.17
pycups==1.9.73
pycurl==7.43.0
pygobject==3.20.0
PyJWT==1.5.0
pymongo==3.4.0
pyparsing==2.0.3
pyquery==1.2.17
PySocks==1.6.7
python-apt==1.1.0b1+ubuntu0.16.4.2
python-dateutil==2.6.1
python-debian==0.1.27
python-systemd==231
pytz==2017.2
pyxdg==0.25
reportlab==3.3.0
requests==2.9.1
scikit-learn==0.19.0
scipy==0.19.1
screen-resolution-extra==0.0.0
sessioninstaller==0.0.0
six==1.10.0
smart-open==1.5.3
stanfordcorenlp==3.9.1.1
stop-words==2015.2.23.1
system-service==0.3
textblob==0.12.0
twilio==6.4.2
ubuntu-drivers-common==0.0.0
ufw==0.35
unattended-upgrades==0.1
unity-scope-calculator==0.1
unity-scope-chromiumbookmarks==0.1
unity-scope-colourlovers==0.1
unity-scope-devhelp==0.1
unity-scope-firefoxbookmarks==0.1
unity-scope-gdrive==0.7
unity-scope-manpages==0.1
unity-scope-openclipart==0.1
unity-scope-texdoc==0.1
unity-scope-tomboy==0.1
unity-scope-virtualbox==0.1
unity-scope-yelp==0.1
unity-scope-zotero==0.1
urllib3==1.13.1
usb-creator==0.3.0
virtualenv==15.0.1
xdiagnose==3.8.4.1
xkit==0.0.0
XlsxWriter==0.7.3
yagmail==0.9.178
